--- App.tsx
@@ -1,53 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { StackNavigationProp } from '@react-navigation/stack';
-import Home from './screens/Home';
-import MapList from './screens/MapList';
-
-const Stack = createNativeStackNavigator();
-
-type RootStackParamList = {
-  Home: undefined, // undefined because you aren't passing any params to the home screen
-  MapList: undefined; 
-};
-
-type ProfileScreenNavigationProp = StackNavigationProp<
-  RootStackParamList,
-  'Home'
->;
-
-type Props = {
-  navigation: ProfileScreenNavigationProp;
-};
-
-
-const HomeScreen = () => {
-  return <Home></Home>
-}
-
-const MapListScreen = () => {
-  return <MapList></MapList>;
-};
-
-export default function App() {
-  return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Home" component={HomeScreen} />
-        <Stack.Screen name="MapList" component={MapListScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-   
-  },
-  
-});
