+++ app.json
@@ -2,6 +2,7 @@
   "expo": {
     "name": "HotMealsOpenSourceExpoRepo",
     "slug": "HotMealsOpenSourceExpoRepo",
+    "entryPoint": "./src/App.tsx",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
@@ -14,9 +15,7 @@
     "updates": {
       "fallbackToCacheTimeout": 0
     },
-    "assetBundlePatterns": [
-      "**/*"
-    ],
+    "assetBundlePatterns": ["**/*"],
     "ios": {
       "supportsTablet": true
     },
